1. Added REST method for getting availability for preparation 2. Added flow for creating/updating receipt solds and change availability at the same time

+++ src/main/java/org/burshtyn/pharmacy/service/preparation/PreparationService.java
@@ -36,8 +36,8 @@ public class PreparationService extends BaseServiceImpl<Preparation, Preparation
         if (optionalPreparation.isPresent()) {
             Preparation preparation = optionalPreparation.get();
 
-            List<Availability> availabilityList = availabilityService.findByPreparation(preparation);
-            for (Availability availability : availabilityList) {
+            Availability availability = availabilityService.findByPreparation(preparation);
+            if (availability != null) {
                 availabilityService.delete(availability);
             }
